# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VSM Mobile",
        "description": "Documentation API VSM PK11",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/dashboard/player": {
            "get": {
                "tags": [
                    "Dashboards"
                ],
                "summary": "Données du Dashboard Joueur",
                "operationId": "d91b4954ce5781c1cbd228839448c6a0",
                "responses": {
                    "200": {
                        "description": "Prochains matchs et état des cotisations du joueur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "next_event": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Match amical vs Bonamoussadi"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "match"
                                                        },
                                                        "event_date": {
                                                            "type": "string",
                                                            "example": "2026-08-25 07:00:00"
                                                        },
                                                        "location": {
                                                            "type": "string",
                                                            "example": "Stade Camrail"
                                                        }
                                                    },
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "attendance_rate": {
                                                    "type": "string",
                                                    "example": "85%"
                                                },
                                                "contribution_status": {
                                                    "type": "string",
                                                    "example": "Up to date"
                                                },
                                                "balance_due": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/coach": {
            "get": {
                "tags": [
                    "Dashboards"
                ],
                "summary": "Données du Dashboard Coach",
                "operationId": "2502ef2f3f06b0aaa5fd979798a7a6bf",
                "responses": {
                    "200": {
                        "description": "Effectif présent et convocations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_players": {
                                                    "type": "integer",
                                                    "example": 24
                                                },
                                                "confirmed_presences": {
                                                    "type": "integer",
                                                    "example": 18
                                                },
                                                "next_session": {
                                                    "type": "string",
                                                    "example": "Entraînement - 2026-08-23 06:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/treasurer": {
            "get": {
                "tags": [
                    "Dashboards"
                ],
                "summary": "Données du Dashboard Trésorier",
                "operationId": "f634b7e3af4275feb09ff6b79f243cc4",
                "responses": {
                    "200": {
                        "description": "Bilan financier du club",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_collected": {
                                                    "type": "integer",
                                                    "example": 450000
                                                },
                                                "pending_payments": {
                                                    "type": "integer",
                                                    "example": 120000
                                                },
                                                "active_contributors": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/admin": {
            "get": {
                "tags": [
                    "Dashboards"
                ],
                "summary": "Données du Dashboard Admin / Président",
                "operationId": "dbfcf7c8a4ae1eecbc58b64accbd2d92",
                "responses": {
                    "200": {
                        "description": "Vue globale du club",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_members": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "active_events": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "fund_balance": {
                                                    "type": "integer",
                                                    "example": 450000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Entrez votre jeton Sanctum sous la forme: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}